Fix triggered RELEASE_ASSERT in ResourceLoader::didReceiveResponse()
This was regressed in https://codereview.chromium.org/1751203003, which removed an early-clear of ResourceLoader::m_loader in ResourceLoader::cancel().

If m_resource->error() triggers a JS event handler that contains an alert(), we will run a nested event loop. Normally, we call setDefersLoading() on all ResourceLoaders to ensure no resource loading messages are processed. However, before calling m_resource->error(), we call m_fetcher->didFailLoading(), which removes the ResourceLoader from ResourceFetcher::m_loaders, which means ResourceFetcher has no way to call setDefersLoading() on it. When we enter the nested event loop, the response IPC for the ResourceLoader that is being cancelled is processed, and we call didReceiveResponse() in an invalid state.

Add back the early clear of ResourceLoader::m_loader to ensure it can't send messages while completing the cancellation, and add a comment to document why it's necessary.

BUG=594072

Review URL: https://codereview.chromium.org/1802693002

Cr-Original-Commit-Position: refs/heads/master@{#381086}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c9b6287c3a354def3608e67621eb609612eb7e5d