[iOS] Defaults to use system libjpeg on iOS.
It is currently not possible to build libjpeg_turbo on iOS (and may
not be a good idea in term of application size) so default to using
the version of libjpeg shipped with the system.

Move the "libs" assignment from the "jpeg" to "system_libjpeg_config"
config.

BUG=459705

Review URL: https://codereview.chromium.org/1806503002

Cr-Original-Commit-Position: refs/heads/master@{#381432}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 689dc0e45d16789f3ce3241d3033601af395c313