Fix textpath is not displayed on use element
If textpath on use element that don't has path on Treescope,
the content of <textPath> element is not displayed.
This reason is that search <path> element on the shadow root.
Because use Treescope of the real element for the id lookup.

BUG=505546

Review URL: https://codereview.chromium.org/1792673005

Cr-Original-Commit-Position: refs/heads/master@{#381447}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a5e8a760a9457ad814990c5cad81c6dba79d141f