Scale the deltaX/deltaY on the mouse wheel events to be consistent with the compositor implementation.
As the page is zoomed in the scroll amount shouldn't increase. This issue
was discovered with the conversion to mouse wheel gestures which uses
gesture scrolls which are scaled; and the layout test was failing for those
cases.

BUG=595441

Review URL: https://codereview.chromium.org/1812653002

Cr-Original-Commit-Position: refs/heads/master@{#381702}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: aa27c866cdef570e9788f9a93a3a3d87aa36ebce