Don't test implementation details for invalid Link headers.
This changes the basic SingleLinkHeaderTest to not verify implementation
details of how LinkHeader deals with invalid headers. If a header is invalid
it doesn't matter what it's url, rel and other attributes are set to.

Also adds a missing EXPECT to compare the expected "as" attribute with the
returned "as" attribute.

Review URL: https://codereview.chromium.org/1778413005

Cr-Original-Commit-Position: refs/heads/master@{#381790}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: bd9d199b6daf57b7b304c5cd16d6a53cba8582f5