WTF: Implement move semantics for values of {List,Linked}HashSet.
This is similar to two earlier patches that implemented move semantics
in HashMap and HashSet:

https://codereview.chromium.org/1764973002/
https://codereview.chromium.org/1798913002/

This patch is for ListHashSet and LinkedHashSet. The changes in this
patch are mostly the same as the previous patches: taking an incoming
value as a universal reference (T&&), and forwarding it as appropriate.

BUG=567139, 582349

Review URL: https://codereview.chromium.org/1807153002

Cr-Original-Commit-Position: refs/heads/master@{#381918}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b1d3886202efe86f89e711a3b61ab250fadffbd2