Fix the check for divide-by-zero in background image tiling
A recent patch changed the width used to find the amount of space
when background-repeat is set to space, but did not update the
corresponding zero check. This patch fixes it and adds a test
for zero sized tiles in background painting.

R=leviw@chromium.org
BUG=595141, 594915

Review URL: https://codereview.chromium.org/1812893002

Cr-Original-Commit-Position: refs/heads/master@{#382058}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d146f1d1b21205841f8982b4fafc87b02b07aa29