media: Improve MediaKeySession::keysStatusesChange() logging.
Currently the log is "MediaKeySession::keysStatusesChange with 3 keys
and usable key: 1", which could mislead people to think that there's
one usable key while "1" actually is a boolean.

With this CL the logging is "MediaKeySession::keysStatusesChange with
3 keys and hasAdditionalUsableKey is true".

In the future we may want to dump the key status for all key IDs which
would be useful for debugging.

BUG=432671

Review URL: https://codereview.chromium.org/1816173002

Cr-Original-Commit-Position: refs/heads/master@{#382409}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b81e7abbe7fb59be9a8803f569e2cadb7022f79f