Close USB device connections when the page is hidden.
This is a temporary measure to make up for the lack of a visible
indication that a page has an open connection to a USB device. If the
page becomes hidden then all open device connections are closed and
cannot be reopened until it is visible again.

BUG=595375

Review URL: https://codereview.chromium.org/1820173002

Cr-Original-Commit-Position: refs/heads/master@{#382654}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d83aa53f1beb27037f5f77fc06d5b6b0c2e05642