Make multipart/x-mixed-replace parser more robust
1. The parser eats leading CR, LF, CRLF and LFLF. It would be better to
eat only LF and CRLF, as other parts such as
net::HttpUtil::LocateEndOfHeaders does the same thing.

2. The parser eats leading EOLs repeatedly. For example, it eats "\n"
when "\n\nfoo" is given, but it eats "\n" twice when "\n" is given
and then "\nfoo" is given.

3. The parser sends the remainder inputs to the client as body data
if it is parsing body and the input ends with LF, although the parser
might be ought to ignore that LF in the future if it's the end-of-body.

BUG=570608

Review URL: https://codereview.chromium.org/1808353003

Cr-Original-Commit-Position: refs/heads/master@{#382790}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 16099ee7cc1e68d6ffea2b50d4faa1763e33a6af