Replace invalid character check with an assertion
In [1] we added a workaround for invalid characters to prevent a
crash. This was fixed in [2] so this patch replaces the workaround
with an assertion that only valid ranges are returned from
SVGTextMetricsCalculator::updateSubrunRangesForCurrentPosition. The
test added in [1] has not been removed because it's one of the few
complicated complex (ha) text tests we have for SVG.

[1] https://crrev.com/44a62c2cb9fa76ebae1aac0845dd9ed509b08e09
[2] https://crrev.com/3f51d6256864b298b15feb93a57c6d986474245a

BUG=595393

Review URL: https://codereview.chromium.org/1829503002

Cr-Original-Commit-Position: refs/heads/master@{#382849}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b4a7f24ba79baa5a099f9c51071a69b65590a065