Fix missing fields in GestureScrollBegin/Update/End event conversions.
Some of the fields were missing from the event conversion path that
occurs when a plugin is used. This caused fields to be populated
with the default value. As a result the default values enabled
overscroll scrolling when using a low precision delta
mouse wheel with an MacOSX machine.

The bug did not indicate which precision of mouse was being
used but overscrolling for PDF files with a touchpad or a magic
mouse (which have high precision deltas) still enable
elastic overscroll; but that is consistent with MacOSX.
This is a behavioural change for MacOSX but is consistent
with Safari when using a high precision mouse.

BUG=596778

Review URL: https://codereview.chromium.org/1829743003

Cr-Original-Commit-Position: refs/heads/master@{#382901}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c1d91e8f7cd46f1cb68ec09e49d4de5c90175d6c