Worker: Move an assertion to a correct place
This CL fixes an assertion failure happening only on non-oilpan builds.

WorkerThread::performShutdownTask() assumes that the count of references to
WorkerGlobalScope is only one. However, WorkerThreadableLoader that has a
reference to the context is still alive at the point and the assertion fails.
This CL moves the assertion to after the line to dispose the loader.

<More details>

The timing to release a reference to the context owned by the loader was changed
by this patch: https://codereview.chromium.org/1749633002

Before the patch, GlobalFetchImpl::StopDetector was an ActiveDOMObject and was
supposed to release the reference before performShutdownTask() as follows:

  WorkerThread::shutdown()
    WorkerGlobalScope::dispose()
      stopActiveDOMObjects();
        StopDetector::stop()
          FetchManager::stop()
            FetchManager::Loader::dispose()
              // Release WorkerThreadableLoader.
  WorkerThread::performShutdownTask()
    // The assertion can pass.

The patch removed StopDetector and instead made FetchManager::Loader to observe
lifecycle of the context to release the reference. The context destruction is
notified after the assertion as follows:

  WorkerThread::shutdown()
  WorkerThread::performShutdownTask()
    // The assertion cannot pass.
    WorkerGlobalScope::notifyContextDestroyed()
      LifecycleNotifier::notifyContextDestroyed()
        FetchManager::contextDestroyed()
          FetchManager::Loader::dispose()
            // Release WorkerThreadableLoader.

BUG=594230

Review URL: https://codereview.chromium.org/1831743002

Cr-Original-Commit-Position: refs/heads/master@{#383048}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0fdc4685d3b7e5134109b814eb08d8c265cdb84c