WTF: Fix Deque::swap for deques with inline capacity.
This patch is a (revised) reland of Erik Corry's earlier patch authored
more than a year ago:
https://codereview.chromium.org/390983010

This fixes the same issue as the original patch tried to do, with a
completely different algorithm, in a hope that my version is more
readable and understandable.

One of the major differences is the use of a "hole" instead of using 0,
1 or 2 ranges. A Deque is basically a ring buffer, where allocations may
happen incontiguously. My patch represents this model as a range with a
hole: for example, if the backing buffer looks like XXXX........XXXX
where "X" is an allocated slot and "." is unallocated, the sequence of
dots is a hole. This is simpler than Erik's original representation,
which is a list of occupied ranges (in this case, two sequences of Xs).

Tests are adopted from Erik's patch.

BUG=360572

Review URL: https://codereview.chromium.org/1829003002

Cr-Original-Commit-Position: refs/heads/master@{#383669}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 015528e62cf9bdc82f51b7af9cac51d18ae86b0b