[css-grid] Do not use content box as non-auto minimum height
When computing the min-height of items with non-auto minimum height we are
incorrectly returning the content box, and thus incorrectly ignoring the
border and padding of the item.

We are actually ignoring the margins too, but since that also affects
widths, we're leaving it for a follow up patch.

Review URL: https://codereview.chromium.org/1834143002

Cr-Original-Commit-Position: refs/heads/master@{#383944}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 68273cbe7ac490816631a8c6339ff1077216f325