Clean up StyleBuilder application of 'content' property
This patch cleans up the application of the 'content' property. Instead
of having a flag to keep track of whether the value is the first in the
list, we can just null-check the m_content variable (well actually we
already have the check...).

Review URL: https://codereview.chromium.org/1839063007

Cr-Original-Commit-Position: refs/heads/master@{#384215}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e0e8abdbb4680687ee482955b8112890a2ea3ed9