Add testing for subscription from service workers
Testing covers subscription from workers both with and without keys, and in
situations where it triggers a read from a manifest or a pre-subscribed worker.

This also cleaned up the unregistration, so that any registrations are searched
for, instead of using a variable to store the registration.

BUG=596471

Review URL: https://codereview.chromium.org/1816123002

Cr-Original-Commit-Position: refs/heads/master@{#384332}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 2a019912946921d445189c61af9c1ab5c3cef860